Define :initialize method for CashRegister class
- Initialize `@inventory_csv` instance variable with `inventory_csv`
- Initialize `@cart_csv` instance variable with `cart_csv`
- Initialize `@inventory` instance variable as an empty array
- Initialize `@cart` instance variable as an empty array
- Call :load_inventory private method to load data from `@inventory_csv`
  into `@inventory` instance variable
- Call :load_cart private method to load data from `@cart_csv` into
  `@cart` instance variable

This commit will add the `initialize` method to the `CashRegister`class.
The method is defined with `inventory_csv` and `cart_csv` parameters
that will be provided during object instantiation and stored into
instance variables.
This commit will also add :load_inventory and :load_cart private methods
in order to provide support to the :initialize method