Define :total_price method for CashRegister class
This commit will add the `total_price` method to the `CashRegister`class.
The method is defined with no parameters. It will define the discounts
for the products, in this specific case it will:

- Initialize `green_tea_promo` variable as an instance of
  BuyOneGetOneFreeDiscount
- Initialize `strawberry_promo` variable as an instance of
  BulkPurchaseDiscount
- Initialize `coffee_promo` variable as an instance of PercentageDiscount

After defining the discounts, it will calculate the total price by applying
the discounts to the specific related products and returning the total