Define :add_product_to_cart method for Controller class
This commit will add the `add_product_to_cart` method to the `Controller`
class. The method takes no arguments. it displays the inventory, let the
user choose the product to add, return the chosen product's index and
add it to the cart, in case the index provided exists. An error message
is displayed if the index provided by the user does exist