Define :remove_product_from_cart method for Controller class
This commit will add the `remove_product_from_cart` method to the
`Controller` class. The method takes no arguments. It displays the
products in the cart, let the user choose the product to remove, returns
the chosen product's index and removes it from the cart, in case the
index provided exists. An error message is displayed if the index
provided by the user does exist