Define :run method for Router class
This commit will add the `run` method to the `Router`class. The method
is defined with no parameters and it will enter a loop where a menu is
displayed to the user, providing a choice, and route the choice to the
corresponding controller's action. The loop will stop when the instance
variable `@running` is set to be false