Modify :load_inventory private method for CashRegister class
This commit will add codes to `load_inventory` private method in the
`CashRegister` class. The method is defined with no parameters and it
will initialize an empty array `codes`. For each line in the csv file,
a condition will be checked: if the array contains the value of `row[0]`,
that will store the code of the product, a message will be displayed and
the iteration will proceed to the next line. Otherwise, a `Product`
instance will be added to the inventory and its code to the `codes` array.
The purpose of this change is to avoid adding products in the inventory
with the same code